% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_func.R
\name{lake_func}
\alias{lake_func}
\title{Lake volume function}
\usage{
lake_func(initial_level, precip_data, outflow = 10, evap = 1)
}
\arguments{
\item{precip_data, }{changing daily precipitation data (gallons)}

\item{outflow}{of 10 gal/day, default = 10}

\item{evap}{(evaporation) of 1 gal/day, default = 1}

\item{intitial_level, }{changing initital lake level (gallons)}
}
\value{
lake_volume (changing volume of the lake) and total_inputs (inputs into the lake)
}
\description{
Lake volume function
}
\author{
Alia Ajina & Sophia Leiker
}
