% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip_results.R
\docType{data}
\name{precip_results}
\alias{precip_results}
\title{Precip Data}
\format{
A data frame with 1 column and 4 rows:
\describe{
  \item{precip_results}{numeric vector in gallons}
}
}
\source{
Random Dataset created by using rnorm(mean = 10, sd = 2, n=nsample)
}
\usage{
precip_results
}
\description{
This dataset is a representation of various precipitation values
}
\keyword{datasets}
